complete db implementation

+++ unidade-3/desafio-3/src/useCases/patient/delete-patient.ts
@@ -1,7 +1,7 @@
 import { PatientRepository } from "../../repositories/patient/PatientRepository";
-import { ScheduleRepository } from "../../repositories/schedule/ScheduleRepository";
-import { DeleteScheduleByUser } from "../schedule/delete-schedule-by-user";
 import { ListPatientFutureSchedule } from "../schedule/list-patient-future-schedule";
+import { ScheduleRepository } from "../../repositories/schedule/ScheduleRepository";
+// import { DeleteScheduleByUser } from "../schedule/delete-schedule-by-user";
 
 export class DeletePatient {
   constructor(
@@ -9,26 +9,33 @@ export class DeletePatient {
     private scheduleRepository: ScheduleRepository
   ) {}
 
-  execute(cpf: string) {
-    const listPatientFutureSchedule = new ListPatientFutureSchedule(
-      this.patientRepository,
+  async execute(cpf: string) {
+    //Não há mais a necessidade de apagar os agendamentos por usuário
+    //pq o relacionamento no BD está com a configuração CASCADE
+
+    const listPatientFutureSchedule = await new ListPatientFutureSchedule(
+      //this.patientRepository,
       this.scheduleRepository
     ).execute(cpf);
+
+    if (!listPatientFutureSchedule) {
+      return "Erro: não foi possível excluir o paciente";
+    }
 
     if (listPatientFutureSchedule.length > 0) {
       return "Erro: paciente está agendado";
     }
 
-    const patient = this.patientRepository.delete(cpf);
+    const patient = await this.patientRepository.delete(cpf);
 
-    if (!patient) {
+    if (patient == 1) {
+      return "Paciente excluído com sucesso";
+    } else {
       return "Erro: não foi possível excluir o paciente";
     }
 
     //Apaga todos os agendamentos futuros do usuário excluído
-    const deleteScheduleByUser = new DeleteScheduleByUser(this.scheduleRepository)
-    deleteScheduleByUser.execute(cpf)
-    
-    return "Paciente excluído com sucesso";
+    // const deleteScheduleByUser = new DeleteScheduleByUser(this.scheduleRepository)
+    // deleteScheduleByUser.execute(cpf)
   }
 }
